+++ src/app/articles/[id]/_components/dropdown-content.jsx
@@ -5,7 +5,7 @@ import { useFormStatus } from "react-dom"
 
 import EllipsisVertical from '@/assets/icons/ic_ellipsis_vertical.svg'
 import { Button } from "@/components/ui/button"
-import { Modal } from "@/components/ui/modal"
+import { Modal } from "@/components/ui/dialog"
 import { deleteComment } from "@/libs/actions"
 // ! server action (Never trust data from the client)
