+++ src/components/ui/dialog/index.js
@@ -1,0 +1,1 @@
+export * from './dialog';
