--- src/components/ui/modal/index.js
@@ -1,1 +1,0 @@
-export * from './modal'
